requirements.txt
pandas>=1.5.0
openpyxl>=3.0.0
rapidfuzz>=2.0.0
PyYAML>=6.0 